Address np deprecation warning where np.array.reshape requires all ints as input

+++ ves/aggregate.py
@@ -30,8 +30,10 @@ def aggregateTable(tableData, rowCount):
 
     # Convert the lists of voltage and current to Numpy arrays
     voltageSpacing = np.array(voltageSpacing[::4], dtype=np.float64)
-    voltage = np.array(voltage, dtype=np.float64).reshape(rowCount / 4, 4)
-    current = np.array(current, dtype=np.float64).reshape(rowCount / 4, 4)
+    voltage = np.array(
+        voltage, dtype=np.float64).reshape(int(rowCount / 4), 4)
+    current = np.array(
+        current, dtype=np.float64).reshape(int(rowCount / 4), 4)
 
     # Take the mean of every four rows from the Qt Table
     meanVoltage = np.mean(voltage, axis=1)
